requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
email-validator==2.1.0
huggingface-hub==0.25.2
python-dotenv
motor==3.3.2
pymongo==4.6.0
beanie==1.24.0
sentence-transformers==2.2.2
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0